+++ src/components/Main/main.scss
