+++ src/components/ProductCard/ProductCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Product } from '../interfaces';
+import { Product } from '../../interfaces';
 
 class ProductCard extends React.Component<Product> {
   constructor(props: Product) {
