+++ src/pages/HomePage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Main from '../components/Main';
+import Main from '../components/Main/Main';
 
 class HomePage extends React.Component {
   render(): React.ReactNode {
